Confirmação de Delete + artigos do próprio usuário

+++ app/Actions/Jetstream/DeleteUser.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Jetstream;
 
 use Laravel\Jetstream\Contracts\DeletesUsers;
+use App\Models\Article;
 
 class DeleteUser implements DeletesUsers
 {
@@ -14,8 +15,22 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
-        $user->deleteProfilePhoto();
-        $user->tokens->each->delete();
-        $user->delete();
+        $hasArticles = false;
+        $allArticles = Article::all();
+
+        foreach ($allArticles as $article){
+            if($article->user->id == $user->id){
+                $hasArticles = true;
+            }
+        }
+
+        if($hasArticles == true){
+            dd('Esta conta não pode ser removida pois possui um artigo ou mais registrado.');
+        } else {
+            $user->deleteProfilePhoto();
+            $user->tokens->each->delete();
+            $user->delete();
+        }
+        
     }
 }
